+++ src/main/java/com/sw1/demo/model/Producto.java
@@ -18,6 +18,6 @@ public class Producto {
     private String id;
     private String nombre;
     private String descripcion;
-    private BigDecimal precio;
+    private Float precio;
     private String stock;
 }
